Reorganização e otimização das rotas, criação do arquivo partials menu-lateral-perfil
