--- controllers/cadastroController.js
@@ -1,7 +1,0 @@
-const controlador = {
-    cadastro: (req, res)=> {
-      res.render ('cadastro')
-    }
-  }
-  
-  module.exports = controlador;
