--- controllers/loginController.js
@@ -1,7 +1,0 @@
-const controlador = {
-    login: (req, res)=> {
-      res.render ('login')
-    }
-  }
-  
-  module.exports = controlador;
