--- controllers/produtosController.js
@@ -1,7 +1,0 @@
-const controlador = {
-    mostraProduto: (req, res)=> {
-      res.render ('produto')
-    }
-  }
-  
-  module.exports = controlador;
