--- routes/cadastro.js
@@ -1,8 +1,0 @@
-var express = require('express');
-var router = express.Router();
-const cadastroController = require('../controllers/cadastroController')
-
-
-router.get('/', cadastroController.cadastro);
-
-module.exports = router;
