+++ routes/indexRouter.js
@@ -4,5 +4,7 @@ const indexController = require('../controllers/indexController')
 
 /* GET home page. */
 router.get('/', indexController.index);
+router.get('/produto', indexController.produto);
+
 
 module.exports = router;
