--- routes/produtos.js
@@ -1,8 +1,0 @@
-var express = require('express');
-var router = express.Router();
-const produtosController = require('../controllers/produtosController')
-
-/* GET home page. */
-router.get('/', produtosController.mostraProduto);
-
-module.exports = router;
