+++ views/carrinho.ejs
@@ -104,7 +104,7 @@
                             </div>
 
                             <div class="caixa_finalizar_pedido_carrinho" >
-                                <a href="/checkout" class="botao_finalizar_pedido_carrinho">finalizar  pedido</a>
+                                <a href="/checkout/pedido-concluido" class="botao_finalizar_pedido_carrinho">finalizar  pedido</a>
                             </div>
