+++ views/login.ejs
@@ -26,7 +26,7 @@
     <main class="main-login">
         <div class="container">
             <div class="div-main-config-padrao">
-                <form action="/perfil" motod="POST">
+                <form action="/usuarios/perfil" motod="POST">
                     <label for="email"> 
                         <span class="material-icons-outlined">email</span>
                          E-mail
@@ -38,7 +38,7 @@
                     </label>
                     <input type="password" name="password" class="password" placeholder="Digite sua senha" required>
                     <button class="botao-login-entrar"  type="submit">Entrar</button>
-                    <a class="botao-login-criar_conta" href="/cadastro">CRIAR CONTA</a>
+                    <a class="botao-login-criar_conta" href="/usuarios/cadastro">CRIAR CONTA</a>
                 </form> 
             </div>
         </div>
