+++ views/partials/header.ejs
@@ -10,7 +10,7 @@
                     <a href="/" class="link-logo-empresa-mobile"><img src="/images/logo2.svg" alt="Imagem Logo TopSales6"></a>
                     <ul class="lista-menu">
                         <li>
-                            <a href="/login"><span class="material-icons">account_circle</span>Minha Conta</a>
+                            <a href="/usuarios/login"><span class="material-icons">account_circle</span>Minha Conta</a>
                         </li>
                         <li>
                             <a href="#"><span class="material-icons-outlined">headset_mic</span>Central de Atendimento</a>
