+++ views/partials/menu-lateral-perfil.ejs
@@ -1,0 +1,16 @@
+<nav class="menu-lateral-area-restrita">                       
+    <ul class="ul-menu-lateral">
+        <li>
+            <a href="/usuarios/perfil">
+                <span class="material-icons">person</span>
+                Minha Conta
+            </a>
+        </li>
+        <li>
+            <a href="/usuarios/meus-dados">
+                <span class="material-icons">article</span>
+                Meus Dados
+            </a>
+        </li>                            
+    </ul>
+</nav>
